requirements.txt
# Fake Job Posting Detection - Requirements

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Machine Learning Libraries
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Natural Language Processing
nltk==3.8.1

# Database
pymongo==4.6.1

# Authentication
firebase-admin==6.5.0

# Environment Variables
python-dotenv==1.0.0

# Optional - For improved performance
joblib==1.3.2
